test(architecture): lock layer boundary tests
Why:
- Issue #211 requires architecture tests to detect layer boundary erosion before production code dependencies drift.

What Changed:
- Split the existing architecture boundary test into project reference, package reference, and source boundary test classes.
- Added shared architecture-test repository helpers for project files, Unity references, source scanning, and public surface declarations.
- Added coverage for Unity asmdef/package allow lists, forbidden namespace references, and internal implementation namespace exposure.

Impact:
- Production code and public runtime behavior are unchanged.
- Unauthorized project references, package references, forbidden namespace usage, or public surface leaks now fail small architecture tests.
